Refactor - Move sign in authentication function to main controller and delete unncessary session routes, controller and view.
