+++ app.js
@@ -7,7 +7,6 @@ const path = require('path');
 const server = require('./app_server/routes/index');
 const user = require('./app_server/routes/user');
 const listing = require('./app_server/routes/listing');
-const userSession = require('./app_server/routes/session');
 const bookingRequest = require('./app_server/routes/bookingRequest');
 const api = require('./app_api/routes/index');
 const apiUser = require('./app_api/routes/user');
@@ -37,7 +36,6 @@ app.use('/', server);
 app.use('/users', user);
 app.use('/listings', listing);
 app.use('/bookingRequests', bookingRequest);
-app.use('/session', userSession);
 app.use('/api', api);
 app.use('/api/users', apiUser);
 app.use('/api/listings', apiListing);
