+++ app_server/views/layout.pug
@@ -9,7 +9,7 @@ html
     header
     block nav
       nav
-       a#sign-in(href="/session/new") Sign In
+       a#sign-in(href="/users") Sign In
        a#sign-up(href="/users/new") Sign Up
     block content
       div#welcome Find your dream place
