+++ app_server/views/users/login.pug
@@ -8,7 +8,7 @@ block nav
     a#sign-up(href="/users/new") Sign Up
 
 block content
-  form(action="/session", method="POST")
+  form(action="/", method="POST")
     fieldset
       legend Welcome Back
       label(for="username") Username
