+++ app_server/views/users/new.pug
@@ -5,7 +5,7 @@ block header
 
 block nav
   nav
-    a#sign-in(href="/session/new") Sign In
+    a#sign-in(href="/users") Sign In
 
 block content
   form(action="/users", method="POST")
